refactor: Estruturacao dos middlewares default do laravel
