+++ app/Http/Middleware/Framework/Authenticate.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Framework;
 
 use Illuminate\Auth\Middleware\Authenticate as Middleware;
