+++ app/Http/Middleware/Framework/EncryptCookies.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Framework;
 
 use Illuminate\Cookie\Middleware\EncryptCookies as Middleware;
