+++ app/Http/Middleware/Framework/ValidateSignature.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Framework;
 
 use Illuminate\Routing\Middleware\ValidateSignature as Middleware;
