+++ app/Http/Middleware/Framework/VerifyCsrfToken.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Middleware;
+namespace App\Http\Middleware\Framework;
 
 use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken as Middleware;
